pyproject.toml
[tool.poetry]
name = "transmission"
version = "1.0.0"
description = "A Django REST API built to proxy Ethereum requests via ShipChain Engine."

license = "Apache-2.0"

authors = [
    "Adam Hodges <ahodges@shipchain.io>"
]

readme = 'README.md'  # Markdown files are supported

repository = "https://github.com/ShipChain/transmission"
homepage = "https://transmission.shipchain.io"

keywords = ['transmission', 'shipchain', 'logistics']

[tool.poetry.dependencies]
python = "==3.6.9"

asn1crypto = "==0.24.0"
aws-requests-auth = "~0.4"
boto3 = "~1.9"
bpython = "==0.17.1"
celery = "==4.2.1"
celery_once = "==2.0.0"
channels = "~2.3.1"
channels-redis = "~2.4.1"
cryptography = "~2.5"
django = "==2.2.14"
django-admin-rangefilter = "~0.5"
django-cors-headers = "==3.0.0"
django-dotenv = "==1.4.2"
django-enumfields = "==0.10.0"
django-environ = "==0.4.5"
django-extensions = "==2.0.6"
django-fancy-cache = "~1.0"
django-fieldsignals = "~0.4"
django-filter = "~2.1"
django-fsm = "~2.6.1"
django-fullurl= "==0.5"
django-influxdb-tagged-metrics = '==1.3.3'
django-redis = "==4.10.0"
django-simple-history = "~2.7.0"
djangorestframework = "==3.10.3"
djangorestframework-gis = "==0.14"
djangorestframework-jsonapi = "~3.0"
djangorestframework-simplejwt = {git = "git://github.com/davesque/django-rest-framework-simplejwt.git", rev = "04a4f7f2e045ac4a49542b7e98350e66431adcf5"}
geocoder = "==1.38.1"
inflection = ">=0.3.1"
influxdb = "==5.0.0"
logstash_formatter = "~0.5"
mapbox = "==0.16.2"
psycopg2 = ">=2.5.4"
pyjwt = "~1.7"
python-jose = "~3.0"
pyyaml = ">=4.2b4"
redis = "~3.3"
requests = "~2.21"
shipchain-common = "~=1.0.28"
uwsgi = "~2.0.17"
watchdog = "~0.9"
whitenoise = "~4.1"
watchtower = "==0.5.5"


[tool.poetry.dev-dependencies]
asynctest = "~0.13"
bandit = "~1.6"
coverage = "==4.5.1"
django-mock-queries = "~2.1"
fpdf = "==1.7.2"
freezegun = "~0.3"
httpretty = "~=0.9"
minio = "==4.0.6"
mock = "==2.0.0"
moto = "^1.3.10"
pillow = "~7"
prospector = "==1.3.0"
pyexcel = "~0.5"
pyexcel-xls = "~0.5"
pyexcel-xlsx = "~0.5"
pytest-asyncio = "~0.14"
pytest-cov = "~2.6"
pytest-django = "~3.4"
pytest-mock = "^1.10"
responses = "==0.9.0"
safety = "^1.8"
websocket-client = "==0.53.0"
